Implement alphabetization checking for entries
I *think* this works, but it has a quirk: 

"-" inside an entry name is ignored. This places entries like [X-Cleaner] as "XC" and not "X-", which sorts differently than we currently do. 

It will not tell you what line the entry is on, but it will tell you which entry it belongs after