Update winapp2ool
New: Merge Modes - Merge now has 2 behaviors for dealing with duplicate section names in the two ini files it's merging. The default is Replace & Add, which will replace any sections in the first ini with the section of the same name from the second ini. The other mode will remove any sections with the same names. Enjoy! 

New: Preliminary support for command line arguments. I'll post a list of these somewhere that isn't this commit message soon. Added basic support for loading winapp2ool with arguments to automate certain tasks. This feature is not fool proof yet (ie. if you try to break it, it will probably break) and for lacking of testing, not necessarily use proof yet either. For this functionality to work, most of the initialization for the modules was rewritten, so most of my testing so far has been to make sure the modules still work for testing the automation (though if the module works then in theory handing it valid command line args should work just as well) 

CCiniDebug was rewritten to a degree. The prompt to save ccleaner.ini has been removed, and it instead automatically saves its changes back to ccleaner.ini.

The prompt to save diff.txt has been removed. It is now available as a menu level toggle instead.