Update Archived entries.ini
Sorry I had used the Debug Tool and forgot to change back the Header.


As for the entries, which have little to less impact, my statement is following:

Downloads Folder Cleaning is supported in Windows Clean Manager also, so it's alright to add it here.
Downloads Part File Cleaning is even very safe to use in most cases.
Downloads Exe and Msi Cleaning is even more safe to use, since most of them are Installer files in General.

As for Audio, Image and Video file type searcher, they should not be added to Winapp3.ini, since they are only for specific searches, which some power users might use and this would help most of the users not to have to write the entry themself.

As for the Infile Cleaning Rules, this is made from me to start a debate in the Community to support Bleachbit, SystemNinja to support these in future, so we can clean the most private entries. The last resort for me was to add them here, so in future developers copy and paste them into their own winapp2.ini files. Perhaps CCleaner could in future support also infile cleaning, which is the most important part.